Add support for TLSDESC
TLSDESC is an alternate method of TLS access. It is part of the ELF TLS
ABI, and is already commonly used in other toolchains. LLVM supports
this for most ELF backends (X86_64, Aarch64, RISC-V). It has always been
the default for Aarch64, but support for RISC-V and X86_64 were added
before the last LLVM release.

More information on TLSDESC can be found in:
https://android.googlesource.com/platform/bionic/+/HEAD/docs/elf-tls.md
or the original RFC:
https://www.fsfla.org/~lxoliva/writeups/TLS/RFC-TLSDESC-ARM.txt

This patch adds a new unstable option `-Z tls-dialect`, which matches
the common `-mtls-dialect=` option used in Clang, GCC, and other
toolchains. This option only needs to be passed through to LLVM to
update the code generation. For targets like Aarch64 that always use
TLSDESC on RISC-V, this has no effect.

In the future if LLVM changes its defaults for TLS to enable TLSDESC,
users who do not wish to change can use the new option to opt out of
TLSDESC.